POC to rewrite parts to typescript
