+++ app.ts
@@ -4,12 +4,13 @@ const conf = require("./conf");
 
 require("fruster-health").start();
 
-authService
+export default authService
 	.start(conf.bus, conf.mongoUrl)
 	.then(function () {
 		log.info("Auth service started and connected to bus", conf.bus);
 	})
-	.catch(function (err) {
+	.catch(function (err: Error) {
 		log.error("Failed starting auth service", err);
 		process.exit(1);
 	});
+
