+++ package.json
@@ -3,8 +3,9 @@
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
-		"start": "node ./app.js",
-		"test": "nyc --reporter=html --check-coverage --lines 90 node ./spec/support/jasmine-runner.js"
+		"dev": "ts-node-dev --respawn --transpileOnly ./app.ts",
+		"start": "ts-node -T ./app.js",
+		"test": "nyc --reporter=html --check-coverage --lines 90 ts-node ./spec/support/jasmine-runner.js"
 	},
 	"dependencies": {
 		"crypto": "^1.0.1",
@@ -25,7 +26,10 @@
 		"fruster-test-utils": "^0.4.0",
 		"jasmine": "^3.3.1",
 		"jasmine-spec-reporter": "^4.2.1",
-		"nyc": "^13.1.0"
+		"nyc": "^13.1.0",
+		"ts-node": "^8.4.1",
+		"ts-node-dev": "^1.0.0-pre.43",
+		"typescript": "^3.6.3"
 	},
 	"engines": {
 		"node": "8.1.2"
